Advertising Sales Channel Prediction
Sales Channel Prediction Case Study  

When a company enters a market, the distribution strategy and channel it uses are keys to its success in the market, as well as market know-how and customer knowledge and understanding. Because an effective distribution strategy under efficient supply-chain management opens doors for attaining competitive advantage and strong brand equity in the market, it is a component of the marketing mix that cannot be ignored. 

The distribution strategy and the channel design have to be right the first time. The case study of the Sales channel includes a detailed study of TV, radio, and newspaper channels. The predict the total sales generated from all the sales channels.